Extract SwiftLanguageVersion into separate file
Reimplement SwiftLanguageVersion and ToolsVersion comparison using public properties
